Use halfs in post-process

+++ Engine/Shaders/Common/Layer.h
@@ -7,7 +7,7 @@
 #include <simd/simd.h>
 #include <metal_stdlib>
 
-float3 add(float3 a, float3 b, float w);
-float3 screen(float3 a, float3 b, float w);
-float3 overlay(float3 a, float3 b, float w);
-float3 softLight(float3 a, float3 b, float w);
+half3 add(half3 a, half3 b, half w);
+half3 screen(half3 a, half3 b, half w);
+half3 overlay(half3 a, half3 b, half w);
+half3 softLight(half3 a, half3 b, half w);
